reverting some chages
the  changes did in this commit 
https://github.com/CESNET/pakiti-client/commit/2a3a092ebd2f2faa3bc4653188fc215a80410999

makes the client reporting an error now when using the pkgdb at cern, while the server correctly stores the report.

It seems that the server does not report  back the "OK" string if we don't pass the --show-error option. Infact the client does not complain if i run the pakiti-client  with -d option.

Could be that the problem is server side but if those changes are not essential i would prefer to revert them.