Review the communication protocol.
Based on issue #1 it was found out that different servers use different ways
to pass the return status (an HTTP code vs. an application string in the
HTTP body).  In order to unify the handling, the client relies solely on
HTTP codes now. In particular the client doesn't expect any more a string
returned by the server (see the --expect option).

The basics of the protocol are described in the documentation.

The change also introduces the Perl LWP library to handle HTTP messaging and
doesn't require any external commands for that. Some command line options
become superfluous with the change and removed.